+++ examples/scopes/data_source.tf
@@ -1,0 +1,17 @@
+provider "oneview" {
+  ov_username   = var.username
+  ov_password   = var.password
+  ov_endpoint   = var.endpoint
+  ov_sslverify  = var.ssl_enabled
+  ov_apiversion = 2600
+  ov_ifmatch    = "*"
+}
+
+# Testing Data Source
+data "oneview_scope" "scope" {
+  name = "Auto-Scope"
+}
+
+#Importing Existing resource
+#resource "oneview_scope" "import_scope"{
+#}